requirements.txt
zenml>=0.93.0
pandas
numpy
scikit-learn
plotly
pygithub
pyyaml
pydantic